% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_level_medians.R
\name{admin_level_medians}
\alias{admin_level_medians}
\title{Compute Median Prices per Administrative Level}
\usage{
admin_level_medians(df, admin_level_col)
}
\arguments{
\item{df}{Monthly JMMI clean dataset}

\item{admin_level_col}{a column name indicating a Libyan Administrative Level}
}
\value{
A dataframe with median prices for each item in the MEB per Administrative Level
}
\description{
Compute Median Prices per Administrative Level
}
