% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_cleaning_log.R
\name{apply_cleaning_log}
\alias{apply_cleaning_log}
\title{Apply a cleaning log to a data frame}
\usage{
apply_cleaning_log(
  df,
  df_values_col,
  cleaning_log_df,
  cleaning_log_values_col,
  by = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Apply a cleaning log to a data frame
}
