% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_data.R
\name{augment_data}
\alias{augment_data}
\title{Apply Modifications from a Logbook and a Deletion Log to Raw Data}
\usage{
augment_data(raw_data, augmented_logbook, augmented_deletion_log)
}
\arguments{
\item{raw_data}{Data frame of the raw JMMI Survey Data}

\item{augmented_logbook}{Data Frame of the Logbook}

\item{augmented_deletion_log}{Data Frame of the Deletion Log}
}
\value{
Data Frame of the Data with the Logbook and Deletion Log Applied
}
\description{
Apply Modifications from a Logbook and a Deletion Log to Raw Data
}
