% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_data_extract.R
\name{augment_data_extract}
\alias{augment_data_extract}
\title{Generate a Data Extract from Raw Data}
\usage{
augment_data_extract(raw_data)
}
\arguments{
\item{raw_data}{Data Frame of the Raw JMMI Survey Data}
}
\value{
Data Frame with two columns for the submission UUIDs and Enumerator IDs
}
\description{
Generate a Data Extract from Raw Data
}
