% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_logbook.R
\name{augment_logbook}
\alias{augment_logbook}
\title{Generate A Logbook from Raw Data}
\usage{
augment_logbook(raw_data, logbook)
}
\arguments{
\item{raw_data}{Data Frame of the Raw JMMI Survey Data}

\item{logbook}{State of the \verb{02_Logook} Sheet before running the function}
}
\value{
Data Frame Listing the changes entered manually to the \verb{02_Logook} Sheet
}
\description{
Generate A Logbook from Raw Data
}
