% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_summary.R
\name{augment_summary}
\alias{augment_summary}
\title{Generate a Data Summary from Raw Data}
\usage{
augment_summary(raw_data, summary, augmented_logbook)
}
\arguments{
\item{raw_data}{Data Frame of the Raw JMMI Survey Data}

\item{summary}{The current state of the \verb{00_Summary} Sheet in the Cleaning Log}

\item{augmented_logbook}{A Data Frame of the Augmented Logbook}
}
\value{
Data Frame Retaining summarising the operations of the cleaning log
}
\description{
This function creates the summary from the raw data if it is empty to begin with.
}
