% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_meb_cost.R
\name{compute_meb_cost}
\alias{compute_meb_cost}
\title{Compute the Cost of the MEB in Libya}
\usage{
compute_meb_cost(df, admin_level_col, item_group = NULL)
}
\arguments{
\item{df}{Monthly JMMI clean dataset}

\item{admin_level_col}{a column name indicating a Libyan Administrative Level}

\item{item_group}{Optional filter to limit the computation to a specific item group in the Minimum Expenditure Basket}
}
\value{
A dataframe of the cost of MEB per Administrative Level
}
\description{
Compute the Cost of the MEB in Libya
}
\examples{
compute_meb_cost(jmmi_2022_feb, q_region, "food")
compute_meb_cost(jmmi_2022_feb, q_district, "hygiene")
compute_meb_cost(jmmi_2022_feb, q_municipality)
}
