% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Identify Outliers while ignoring those mentioned in a logbook}
\usage{
find_outliers(df, logbook)
}
\arguments{
\item{df}{Data Frame of the Raw JMMI Survey Data}

\item{logbook}{Data Frame of the \verb{02_Logbook} sheet}
}
\value{
Data Frame
}
\description{
Identify Outliers while ignoring those mentioned in a logbook
}
