% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_enumerator_id.R
\name{generate_enumerator_id}
\alias{generate_enumerator_id}
\title{Generate a Unique ID for each enumerator}
\usage{
generate_enumerator_id(df)
}
\arguments{
\item{df}{Data Frame of the Raw JMMI Survey Data}
}
\value{
Data Frame of Submissions UUIDs and Unique IDs for the corresponding enumerators
}
\description{
Generate a Unique ID for each enumerator
}
