% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_columns.R
\name{get_item_columns}
\alias{get_item_columns}
\title{Get item columns from survey}
\usage{
get_item_columns(df, col_name)
}
\arguments{
\item{col_name}{}
}
\description{
Get item columns from survey
}
