% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_data_collection.R
\name{init_data_collection}
\alias{init_data_collection}
\title{Initialize a JMMI Project}
\usage{
init_data_collection(period, base_path)
}
\arguments{
\item{period}{Character Vector in the Format \code{JMMI_2023_01}}

\item{base_path}{Path to the project}
}
\description{
Initialize a JMMI Project
}
