% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmmi_2022_jan}
\alias{jmmi_2022_jan}
\title{JMMI dataset for January 2022}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 810 rows and 462 columns.
}
\source{
\url{https://www.impact-repository.org/document/reach/cd36ef9f/reach_lby_dataset_joint_market_monitoring_initiative_jmmi_January_2022.xlsx}
}
\usage{
jmmi_2022_jan
}
\description{
The clean dataset for January 2022 JMMI
}
\keyword{datasets}
