% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_factory.R
\name{kobo_download_target}
\alias{kobo_download_target}
\title{Download Data From KoBo ToolBox}
\usage{
kobo_download_target(
  period,
  uid,
  kobo_username = Sys.getenv("KOBO_USERNAME"),
  kobo_password = Sys.getenv("KOBO_PASSWORD"),
  base_path = here::here(),
  raw_data_age = as.difftime(8, units = "hours")
)
}
\arguments{
\item{period}{a character vector in the format "JMMI_YEAR_MONTH"}

\item{uid}{the uid of the form}

\item{kobo_username}{KoBo Username, defaults to the environment variable KOBO_USERNAME}

\item{kobo_password}{KoBo Password, defaults to the environment variable KOBO_PASSWORD}

\item{base_path}{path where the file will be downloaded}

\item{raw_data_age}{the duration after which the download will be refreshed}
}
\description{
Download Data From KoBo ToolBox
}
