% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meb_weights}
\alias{meb_weights}
\title{MEB weights for the Libyan JMMI}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 45 rows and 4 columns.
}
\usage{
meb_weights
}
\description{
Weights of MEB items used for the cost of the MEB
in Libya
}
\details{
\describe{
\item{item}{Item label as per the standard KoBo tool used for data collection}
\item{group}{The item group of the item}
\item{label}{A readable label for the item, including the unit}
\item{weight}{The weight contribution of the item towards the final cost of MEB}
}
}
\keyword{datasets}
