% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_factory.R
\name{pipeline}
\alias{pipeline}
\title{Data Cleaning Pipeline}
\usage{
pipeline(
  period,
  kobo_username = Sys.getenv("KOBO_USERNAME"),
  kobo_password = Sys.getenv("KOBO_PASSWORD"),
  uid,
  base_path,
  raw_data_age = as.difftime(8, units = "hours")
)
}
\arguments{
\item{raw_data_age}{}
}
\description{
Data Cleaning Pipeline
}
