% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtb-package.R
\docType{package}
\name{qtb-package}
\alias{qtb}
\alias{qtb-package}
\title{qtb: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
