% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtb_visnetwork.R
\name{qtb_visnetwork}
\alias{qtb_visnetwork}
\title{visNetwork dependency graph for \code{period}}
\usage{
qtb_visnetwork(period, ...)
}
\arguments{
\item{period}{a string period id}

\item{...}{arguments passed down to \link[targets]{tar_visnetwork}}
}
\value{
A visNetwork HTML widget object
}
\description{
Visualise the pipeline dependency graph for the specified data collection period
}
