% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracked_municipalities}
\alias{tracked_municipalities}
\title{Tracked Municipalities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41 rows and 2 columns.
}
\usage{
tracked_municipalities
}
\description{
The list of municipalities where data collection
usually occurs.
}
\details{
\describe{
\item{municipality_name_en}{The Municipality's name in English}
\item{position}{A helper attribute indicating the position of the Municipality in table reports}
}
}
\keyword{datasets}
