% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cleaning_log.R
\name{write_cleaning_log}
\alias{write_cleaning_log}
\title{Export a Cleaning Log of Augmented data}
\usage{
write_cleaning_log(
  wb,
  raw_data,
  clean_data,
  data_extract,
  logbook,
  summary,
  base_path
)
}
\arguments{
\item{wb}{Workbook of the Template Cleaning Log}

\item{raw_data}{Data Frame of the Raw JMMI Survey Data}

\item{clean_data}{Data Frame of the Augmented Raw Data}

\item{data_extract}{Data Frame of the Augmented Data Extract}

\item{logbook}{Data Frame of the Augmented Logbook}

\item{summary}{Data Frame of the Augmented Summary}

\item{base_path}{Path to the export directory}
}
\description{
Export a Cleaning Log of Augmented data
}
