% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_level_medians.R
\name{admin_level_medians}
\alias{admin_level_medians}
\title{Compute Median Prices per Administrative Level}
\usage{
admin_level_medians(df, admin_level_col, item_group = NULL)
}
\arguments{
\item{df}{Monthly JMMI clean dataset}

\item{admin_level_col}{a column name indicating a Libyan Administrative Level}

\item{item_group}{Optional filter to limit the computation to a specific item group in the Minimum Expenditure Basket}
}
\value{
A dataframe with median prices for each item in the MEB per Administrative Level
}
\description{
Compute Median Prices per Administrative Level
}
\examples{
admin_level_medians(jmmi_2022_feb, q_region)
admin_level_medians(jmmi_2022_feb, q_district)
admin_level_medians(jmmi_2022_feb, q_municipality)
}
