% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_level_medians.R
\name{admin_level_medians}
\alias{admin_level_medians}
\title{Compute Median Prices per Administrative Level}
\usage{
admin_level_medians(df, admin_level = "municipality")
}
\arguments{
\item{df}{Monthly JMMI clean dataset}

\item{admin_level}{an administrative level in Libya, can be one of "municipality", "district", "region", or "overall"}
}
\value{
A dataframe with median prices for each item in the MEB per Administrative Level
}
\description{
Compute Median Prices per Administrative Level
}
\examples{
admin_level_medians(jmmi_2022_feb, "region")
admin_level_medians(jmmi_2022_feb, "district")
admin_level_medians(jmmi_2022_feb, "municipality")
}
