% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_meb_cost.R
\name{compute_meb_cost}
\alias{compute_meb_cost}
\title{Compute the Cost of the MEB in Libya}
\usage{
compute_meb_cost(df, admin_level = "municipality", precision = 3)
}
\arguments{
\item{df}{Monthly JMMI clean dataset}

\item{admin_level}{an administrative level in Libya, can be one of "municipality", "district", "region", or "overall"}

\item{precision}{The number of decimal places to keep when rounding results}
}
\value{
A dataframe of the cost of MEB per Administrative Level
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
compute_meb_cost(jmmi_2022_feb, "region")
compute_meb_cost(jmmi_2022_feb, "district")
compute_meb_cost(jmmi_2022_feb, "municipality")
}
