% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_factory.R
\name{pipeline}
\alias{pipeline}
\title{Data Cleaning Pipeline}
\usage{
pipeline(
  period,
  kobo_username = Sys.getenv("KOBO_USERNAME"),
  kobo_password = Sys.getenv("KOBO_PASSWORD"),
  uid,
  base_path,
  raw_data_age = as.difftime(8, units = "hours")
)
}
\arguments{
\item{period}{Period Name}

\item{kobo_username}{KoBo Username, defaults to KOBO_USERNAME from environment}

\item{kobo_password}{KoBo Password, defaults to KOBO_PASSWORD from environment}

\item{uid}{Form UID on KoBo Toolbox}

\item{base_path}{Path to the project}

\item{raw_data_age}{Duration for refreshing the KoBo Data Export}
}
\description{
Data Cleaning Pipeline
}
