% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_factory.R
\name{pipeline}
\alias{pipeline}
\title{Data Cleaning Pipeline}
\usage{
pipeline(period, base_path)
}
\arguments{
\item{period}{Period Name}

\item{base_path}{Path to the project}
}
\description{
Data Cleaning Pipeline
}
