% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtb-package.R
\docType{package}
\name{qtb-package}
\alias{qtb}
\alias{qtb-package}
\title{qtb: The MEB Toolkit for REACH's Libya Mission}
\description{
A toolkit for computing the cost of a Minimum Expenditure Basket (MEB) under the Joint Market Monitoring Initiative (JMMI) of REACH in Libya.
}
\keyword{internal}
